Move exceptions to the namespace that they are most appropriate for.
