+++ Netimobiledevice/Backup/Mobilebackup2Exception.cs
@@ -1,6 +1,4 @@
-﻿using Netimobiledevice.Exceptions;
-
-namespace Netimobiledevice.Backup
+﻿namespace Netimobiledevice.Backup
 {
     public class Mobilebackup2Exception : NetimobiledeviceException
     {
