+++ Netimobiledevice/DeviceLink/DeviceLinkException.cs
@@ -1,6 +1,4 @@
-﻿using Netimobiledevice.Exceptions;
-
-namespace Netimobiledevice.DeviceLink
+﻿namespace Netimobiledevice.DeviceLink
 {
     public class DeviceLinkException : NetimobiledeviceException
     {
