+++ Netimobiledevice/Diagnostics/DeprecatedException.cs
@@ -1,0 +1,5 @@
+﻿namespace Netimobiledevice.Diagnostics;
+
+public class DeprecatedException(string message) : NetimobiledeviceException(message)
+{
+}
