+++ Netimobiledevice/Diagnostics/DiagnosticsException.cs
@@ -1,6 +1,4 @@
-﻿using Netimobiledevice.Exceptions;
-
-namespace Netimobiledevice.Diagnostics
+﻿namespace Netimobiledevice.Diagnostics
 {
     public class DiagnosticsException : NetimobiledeviceException
     {
