--- Netimobiledevice/Exceptions/ConnectionFailedException.cs
@@ -1,6 +1,0 @@
-﻿namespace Netimobiledevice.Exceptions
-{
-    public class ConnectionFailedException : NetimobiledeviceException
-    {
-    }
-}
