--- Netimobiledevice/Exceptions/DeprecatedException.cs
@@ -1,7 +1,0 @@
-﻿namespace Netimobiledevice.Exceptions
-{
-    public class DeprecatedException : NetimobiledeviceException
-    {
-        public DeprecatedException(string message) : base(message) { }
-    }
-}
