--- Netimobiledevice/Exceptions/DeviceDisconnectedException.cs
@@ -1,10 +1,0 @@
-﻿namespace Netimobiledevice.Exceptions
-{
-    /// <summary>
-    /// Exception thrown when the device is disconnected.
-    /// </summary>
-    public class DeviceDisconnectedException : NetimobiledeviceException
-    {
-        public DeviceDisconnectedException() : base("Device disconnected") { }
-    }
-}
