--- Netimobiledevice/Exceptions/FatalPairingException.cs
@@ -1,9 +1,0 @@
-﻿namespace Netimobiledevice.Exceptions
-{
-    public class FatalPairingException : NetimobiledeviceException
-    {
-        public FatalPairingException() : base() { }
-
-        public FatalPairingException(string message) : base(message) { }
-    }
-}
