--- Netimobiledevice/Exceptions/PasswordRequiredException.cs
@@ -1,7 +1,0 @@
-﻿namespace Netimobiledevice.Exceptions
-{
-    public class PasswordRequiredException : NetimobiledeviceException
-    {
-        public PasswordRequiredException(string message) : base(message) { }
-    }
-}
