--- Netimobiledevice/Exceptions/ServiceStartException.cs
@@ -1,7 +1,0 @@
-﻿namespace Netimobiledevice.Exceptions
-{
-    public class ServiceStartException : NetimobiledeviceException
-    {
-        public ServiceStartException(string message) : base(message) { }
-    }
-}
