+++ Netimobiledevice/Lockdown/FatalPairingException.cs
@@ -1,0 +1,8 @@
+﻿namespace Netimobiledevice.Lockdown;
+
+public class FatalPairingException : NetimobiledeviceException
+{
+    public FatalPairingException() : base() { }
+
+    public FatalPairingException(string message) : base(message) { }
+}
