+++ Netimobiledevice/Lockdown/NoDeviceConnectedException.cs
@@ -1,0 +1,5 @@
+﻿namespace Netimobiledevice.Lockdown;
+
+public class NoDeviceConnectedException : NetimobiledeviceException
+{
+}
