+++ Netimobiledevice/Lockdown/NotPairedException.cs
@@ -1,0 +1,5 @@
+﻿namespace Netimobiledevice.Lockdown;
+
+public class NotPairedException : NetimobiledeviceException
+{
+}
