+++ Netimobiledevice/Lockdown/PasswordRequiredException.cs
@@ -1,0 +1,5 @@
+﻿namespace Netimobiledevice.Lockdown;
+
+public class PasswordRequiredException(string message) : NetimobiledeviceException(message)
+{
+}
