+++ Netimobiledevice/Lockdown/ServiceStartException.cs
@@ -1,0 +1,5 @@
+﻿namespace Netimobiledevice.Lockdown;
+
+public class ServiceStartException(string message) : NetimobiledeviceException(message)
+{
+}
