+++ Netimobiledevice/Misagent/MisagentException.cs
@@ -1,5 +1,3 @@
-﻿using Netimobiledevice.Exceptions;
-
-namespace Netimobiledevice.Misagent;
+﻿namespace Netimobiledevice.Misagent;
 
 public sealed class MisagentException(string message) : NetimobiledeviceException(message) { }
