+++ Netimobiledevice/Usbmuxd/UsbmuxVersionException.cs
@@ -1,5 +1,3 @@
-﻿using Netimobiledevice.Exceptions;
-
-namespace Netimobiledevice.Usbmuxd;
+﻿namespace Netimobiledevice.Usbmuxd;
 
 public class UsbmuxVersionException(string message) : NetimobiledeviceException(message) { }
